Revert "ticket:5 fighting with getbarista"
This reverts commit afd68b6221b45bd0072f4c1d744555938c6568be.

+++ post_deploy.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
 echo "Cleaning old static..."
-rm -rf assets/libs
+if [ ! -d assets/libs ]; then
+    rm -rf assets/libs
+fi
 rm -rf static/*
 echo "Installing bower dependencies..."
 make bower-install
